Preprocess.py creates a new ogr.wkbPolygon object, containing an ogr.wkbLinearRing object, from the 15s Niger River Basin ESRI shapefile. This new object must be created, since apparently the ESRI data source, which contains the polygon and linear ring geometries, is silently removed after the loadmaps() function returns. After the new geometry is created, the data source is explicitly destroyed, and the new ogr.Geometry object is returned along with two python maps: a map from a segment ID to the list of river network segment IDs immediately upstream, and a map from IDs to segments in geographical coordinates, oriented from upstream to downstream. The segments are lists of the form [[x1, y1], [x2, y2]], where [x1, y1] is upstream of [x2,y2]. Not allocating a new geometry led to unpredictable behavior, which alternated between infinite looping and segmentation faults when the basin boundary function was called.
The breadth-first search code bfs.py was modified to compute
convex hulls from multipoint geometries derived from the collection
of points upstream of a given point. We'll call the convex hull
produced by the ConvexHull() method a sub-basin hull. The
sub-basin hull is a geometry containing a single ring. This
is tested against the 15s basin geometry created above. If they
intersect, and if the intersection returns a geometry with
precisely one ring, the ring is returned. Otherwise, the convex
hull is returned. Experiments show that the best results
are obtained by intersecting the basin boundary with a convex
hull derived from Strahler order 1 points upstream of the
given segment ID. Higher Strahler-order derived sub-basin hulls
can produce multiple rings (the reason is unclear), and in that
case the convex hull generally is not entirely contained within
the basin. The result is unaesthetic.